Move to GRUB as bootloader

+++ kern/fs/vfs.c
@@ -116,7 +116,7 @@ int vfs_mount_filesystem(ptable_entry_t* fs, char* mountPoint) {
 		// Add to mountpoint hashmap
 		hashmap_insert(mountPointMap, mountPoint, superblock);
 
-		return superblock;
+		return (int) superblock;
 	} else {
 		kprintf("Unknown filesystem type 0x%X\n", fs->type);
 		return -1;
